requirements.txt
python-dotenv
openai
fastapi
requests
uvicorn
pydantic
gunicorn
tenacity